+++ wayland-client/CHANGELOG.md
@@ -10,6 +10,7 @@
   dispatch implementations
 - `BindError` now includes the requested and available version, or interface name that failed to bind.
 - `GlobalList` now is updated with normal queueing, requires a `GlobalListHandler` implementation
+- Rename `GlobalList::bind` to `bind_singleton`
 
 ## 0.31.14-- 2026-03-30